Fix last few issues

+++ ClassicalSharp/2D/Widgets/HotbarWidget.cs
@@ -69,7 +69,7 @@ namespace ClassicalSharp.Gui.Widgets {
 			drawer.BeginBatch(game, cache.vertices, cache.vb);
 			
 			for (int i = 0; i < hotbarCount; i++) {
-				BlockID block = (byte)game.Inventory.Hotbar[i];
+				BlockID block = game.Inventory.Hotbar[i];
 				int x = (int)(X + barXOffset + (elemSize + borderSize) * i + elemSize / 2);
 				int y = (int)(game.Height - barHeight / 2);
